[JIT] Subgraph utils: add an optional vmap argument to the API to allow retrieving value mappings.
This functionality is needed when we want to not lose track of
nodes/values as we merge and unmerge them into other nodes. For
instance, if we have a side data structure with some meta information
about values or nodes, this new functionality would allow to keep that
metadata up to date after merging and unmerging nodes.

ghstack-source-id: 15bf3a7c47ad4f59885ea0bcfb17c4ee8a09073c
Pull Request resolved: https://github.com/pytorch/pytorch/pull/43235